requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.2.0
idna==3.3
ISS-Info==1.0.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
